+++ src/visualize/+mag/+graphics/+chart/Bubblechart.m
@@ -32,15 +32,13 @@ classdef Bubblechart < mag.graphics.chart.Chart
                 graph (1, :) matlab.graphics.Graphics
             end
 
-            filteredData = this.filterData(data);
-
-            arguments = {this.getXData(filteredData), filteredData{:, this.YVariables}, filteredData{:, this.SVariables}};
+            args = {this.getXData(data), data{:, this.YVariables}, data{:, this.SVariables}};
 
             if ~isempty(this.CVariables)
-                arguments = [arguments, {filteredData{:, this.CVariables}}];
+                args = [args, {data{:, this.CVariables}}];
             end
 
-            graph = bubblechart(axes, arguments{:});
+            graph = bubblechart(axes, args{:});
         end
     end
 end