+++ src/visualize/+mag/+graphics/+chart/Chart.m
@@ -6,8 +6,6 @@ classdef (Abstract) Chart < matlab.mixin.Heterogeneous & mag.mixin.SetGet
         XVariable string {mustBeScalarOrEmpty}
         % YVARIABLES Name of variables plotted on y-axis.
         YVariables (1, :) string
-        % FILTERS Filter x- and y-axis variables.
-        Filters logical = logical.empty()
     end
 
     methods (Abstract)
@@ -17,30 +15,6 @@ classdef (Abstract) Chart < matlab.mixin.Heterogeneous & mag.mixin.SetGet
     end
 
     methods (Access = protected)
-
-        function filteredData = filterData(this, data)
-        % FILTERDATA Filter data based on specified filter.
-
-            if isempty(this.Filters)
-
-                % No filtering.
-                filteredData = data;
-            elseif (size(this.Filters, 2) == 1) && (numel(this.YVariables) >= 1)
-
-                % Same filtering for each y-axis variable.
-                filteredData = data(this.Filters, :);
-            elseif (size(this.Filters, 2) == numel(this.YVariables))
-
-                % Different filtering for each y-axis variable.
-                filteredData = cell(1, numel(this.YVariables));
-
-                for i = 1:numel(this.YVariables)
-                    filteredData{i} = data(this.Filters(:, i), :);
-                end
-            else
-                error("Mismatch between filter and y-axis variable selection.");
-            end
-        end
 
         function xData = getXData(this, data)
         % GETXDATA Retrieve x-axis data, based on selected variable name.