+++ src/visualize/+mag/+graphics/+chart/Histogram.m
@@ -25,10 +25,7 @@ classdef Histogram < mag.graphics.chart.Chart & mag.graphics.mixin.ColorSupport
                 graph (1, :) matlab.graphics.Graphics
             end
 
-            filteredData = this.filterData(data);
-            assert(~iscell(filteredData), "Different filter lengths not currently supported.");
-
-            graph = histogram(axes, filteredData{:, this.YVariables});
+            graph = histogram(axes, data{:, this.YVariables});
 
             this.applyColorStyle(graph, "FaceColor");
         end