+++ resources/ReleaseNotes.md
@@ -6,3 +6,4 @@
 
 - Rename `AutomatedAnalysis` to `IMAPTestingAnalysis`
 - Rename `Result` to `PSD`
+- Remove support for `Filters` in charts